requirements_large.txt
numpy>=1.21.0
Pillow>=9.0.0
tifffile>=2022.2.9